push-notifs/requirements.txt
cffi==1.14.3
cryptography==2.9.2
h2==2.6.2
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
pycparser==2.20
PyJWT==1.7.1
six==1.15.0
